+++ weblate/api/views.py
@@ -803,6 +803,9 @@ class ComponentViewSet(
 ):
     """Translation components API."""
 
+    raw_urls: tuple[str, ...] = "component-file"
+    raw_formats = ("zip", *(f"zip:{exporter}" for exporter in EXPORTERS))
+
     queryset = Component.objects.none()
     serializer_class = ComponentSerializer
     lookup_fields = ("project__slug", "slug")
@@ -1003,8 +1006,8 @@ class ComponentViewSet(
         instance.links.remove(project)
         return Response(status=HTTP_204_NO_CONTENT)
 
-    @action(detail=True, methods=["get"], url_path="file")
-    def download_archive(self, request, **kwargs):
+    @action(detail=True, methods=["get"])
+    def file(self, request, **kwargs):
         # Implementation is analogous to files#download_component, but we can't reuse
         #  that here because the lookup for the component is different
         instance = self.get_object()
@@ -1015,6 +1018,7 @@ class ComponentViewSet(
 
         requested_format = request.query_params.get("format", "zip")
         return download_multi(
+            request,
             instance.translation_set.all(),
             [instance],
             requested_format,